pyproject.toml
[project]
name = "wildfire-nowcast-agent"
version = "0.1.0"
description = "Intelligent wildfire detection and response agent using Amazon Bedrock AgentCore with Strands Agent framework"

readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents",
    "strands-agents-tools",
    "boto3",
    "requests",
    "folium",
    "pillow",
    "matplotlib",
    "seaborn",
    "pandas",
    "numpy",
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit",
    "python-dotenv",
    "owslib",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
